+++ core/snapshot/databases/metadata/in_memory.go
@@ -18,7 +18,7 @@ package metadata
 import (
 	"sort"
 
-	tmtypes "github.com/tendermint/tendermint/abci/types"
+	tmtypes "github.com/cometbft/cometbft/abci/types"
 )
 
 type InMemoryDatabase struct {