+++ core/snapshot/databases/metadata/level_db_test.go
@@ -21,9 +21,9 @@ import (
 	"code.vegaprotocol.io/vega/core/snapshot/databases/metadata"
 	"code.vegaprotocol.io/vega/paths"
 
+	tmtypes "github.com/cometbft/cometbft/abci/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	tmtypes "github.com/tendermint/tendermint/abci/types"
 )
 
 func TestLevelDBDatabase(t *testing.T) {