+++ core/snapshot/tree/tree.go
@@ -26,8 +26,8 @@ import (
 	"code.vegaprotocol.io/vega/logging"
 	snapshotpb "code.vegaprotocol.io/vega/protos/vega/snapshot/v1"
 
+	tmtypes "github.com/cometbft/cometbft/abci/types"
 	"github.com/cosmos/iavl"
-	tmtypes "github.com/tendermint/tendermint/abci/types"
 )
 
 // Tree is a high-level structure that abstract the management of the AVL away