+++ core/spam/simple_spam_policy.go
@@ -20,7 +20,6 @@ import (
 	"errors"
 	"sort"
 	"sync"
-	"time"
 
 	"code.vegaprotocol.io/vega/core/blockchain/abci"
 	"code.vegaprotocol.io/vega/core/types"
@@ -42,13 +41,10 @@ type SimpleSpamPolicy struct {
 	minTokensParamName  string
 	maxAllowedParamName string
 
-	partyToCount          map[string]uint64           // commands that are already on blockchain
-	blockPartyToCount     map[string]uint64           // commands in the current block
-	bannedParties         map[string]int64            // parties banned -> ban end time
-	partyBlockRejects     map[string]*blockRejectInfo // total vs rejection in the current block
-	currentEpochSeq       uint64                      // current epoch sequence
-	lock                  sync.RWMutex                // global lock to sync calls from multiple tendermint threads
-	banErr                func(until time.Time) error
+	partyToCount          map[string]uint64 // commands that are already on blockchain
+	blockPartyToCount     map[string]uint64 // commands in the current block
+	currentEpochSeq       uint64            // current epoch sequence
+	lock                  sync.RWMutex      // global lock to sync calls from multiple tendermint threads
 	insufficientTokensErr error
 	tooManyCommands       error
 }
@@ -56,21 +52,16 @@ type SimpleSpamPolicy struct {
 // NewSimpleSpamPolicy instantiates the simple spam policy.
 func NewSimpleSpamPolicy(policyName string, minTokensParamName string, maxAllowedParamName string, log *logging.Logger, accounts StakingAccounts) *SimpleSpamPolicy {
 	return &SimpleSpamPolicy{
-		log:                 log,
-		accounts:            accounts,
-		policyName:          policyName,
-		partyToCount:        map[string]uint64{},
-		blockPartyToCount:   map[string]uint64{},
-		bannedParties:       map[string]int64{},
-		partyBlockRejects:   map[string]*blockRejectInfo{},
-		lock:                sync.RWMutex{},
-		minTokensParamName:  minTokensParamName,
-		maxAllowedParamName: maxAllowedParamName,
-		minTokensRequired:   num.UintZero(),
-		maxAllowedCommands:  1, // default is allow one per epoch
-		banErr: func(until time.Time) error {
-			return errors.New("party is banned from submitting " + policyName + " until the earlier between " + until.String() + " and the beginning of the next epoch")
-		},
+		log:                   log,
+		accounts:              accounts,
+		policyName:            policyName,
+		partyToCount:          map[string]uint64{},
+		blockPartyToCount:     map[string]uint64{},
+		lock:                  sync.RWMutex{},
+		minTokensParamName:    minTokensParamName,
+		maxAllowedParamName:   maxAllowedParamName,
+		minTokensRequired:     num.UintZero(),
+		maxAllowedCommands:    1, // default is allow one per epoch
 		insufficientTokensErr: errors.New("party has insufficient associated governance tokens in their staking account to submit " + policyName + " request"),
 		tooManyCommands:       errors.New("party has already submitted the maximum number of " + policyName + " requests per epoch"),
 	}
@@ -87,22 +78,11 @@ func (ssp *SimpleSpamPolicy) Serialise() ([]byte, error) {
 
 	sort.SliceStable(partyToCount, func(i, j int) bool { return partyToCount[i].Party < partyToCount[j].Party })
 
-	bannedParties := make([]*types.BannedParty, 0, len(ssp.bannedParties))
-	for party, until := range ssp.bannedParties {
-		bannedParties = append(bannedParties, &types.BannedParty{
-			Party: party,
-			Until: until,
-		})
-	}
-
-	sort.SliceStable(bannedParties, func(i, j int) bool { return bannedParties[i].Party < bannedParties[j].Party })
-
 	payload := types.Payload{
 		Data: &types.PayloadSimpleSpamPolicy{
 			SimpleSpamPolicy: &types.SimpleSpamPolicy{
 				PolicyName:      ssp.policyName,
 				PartyToCount:    partyToCount,
-				BannedParty:     bannedParties,
 				CurrentEpochSeq: ssp.currentEpochSeq,
 			},
 		},
@@ -118,11 +98,6 @@ func (ssp *SimpleSpamPolicy) Deserialise(p *types.Payload) error {
 	for _, ptc := range pl.PartyToCount {
 		ssp.partyToCount[ptc.Party] = ptc.Count
 	}
-	ssp.bannedParties = make(map[string]int64, len(pl.BannedParty))
-	for _, bp := range pl.BannedParty {
-		ssp.bannedParties[bp.Party] = bp.Until
-	}
-
 	ssp.currentEpochSeq = pl.CurrentEpochSeq
 
 	return nil
@@ -158,49 +133,20 @@ func (ssp *SimpleSpamPolicy) Reset(epoch types.Epoch) {
 
 	// reset counts
 	ssp.partyToCount = map[string]uint64{}
-
-	// clear banned on new epoch
-	ssp.bannedParties = map[string]int64{}
-
 	ssp.blockPartyToCount = map[string]uint64{}
-	ssp.partyBlockRejects = map[string]*blockRejectInfo{}
 }
 
-// EndOfBlock is called at the end of the processing of the block to carry over state and trigger bans if necessary.
-func (ssp *SimpleSpamPolicy) EndOfBlock(blockHeight uint64, now time.Time, banDuration time.Duration) {
+func (ssp *SimpleSpamPolicy) UpdateTx(tx abci.Tx) {
 	ssp.lock.Lock()
 	defer ssp.lock.Unlock()
-	// add the block's counters to the epoch's
-	for party, count := range ssp.blockPartyToCount {
-		if _, ok := ssp.partyToCount[party]; !ok {
-			ssp.partyToCount[party] = 0
-		}
-		ssp.partyToCount[party] += count
+	if _, ok := ssp.partyToCount[tx.Party()]; !ok {
+		ssp.partyToCount[tx.Party()] = 0
 	}
-
-	ssp.blockPartyToCount = map[string]uint64{}
-
-	// release bans
-	nowNano := now.UnixNano()
-	for k, v := range ssp.bannedParties {
-		if nowNano >= v {
-			delete(ssp.bannedParties, k)
-		}
-	}
-
-	endBanTime := now.Add(banDuration).UnixNano()
-
-	// ban parties with more than <banFactor> rejection rate in the block
-	for p, bStats := range ssp.partyBlockRejects {
-		if num.DecimalFromInt64(int64(bStats.rejected)).Div(num.DecimalFromInt64(int64(bStats.total))).GreaterThanOrEqual(banFactor) {
-			ssp.bannedParties[p] = endBanTime
-		}
-	}
-	ssp.partyBlockRejects = map[string]*blockRejectInfo{}
+	ssp.partyToCount[tx.Party()]++
 }
 
-// PostBlockAccept is called to verify a transaction from the block before passed to the application layer.
-func (ssp *SimpleSpamPolicy) PostBlockAccept(tx abci.Tx) (bool, error) {
+// CheckBlockTx is called to verify a transaction from the block before passed to the application layer.
+func (ssp *SimpleSpamPolicy) CheckBlockTx(tx abci.Tx) error {
 	party := tx.Party()
 
 	ssp.lock.Lock()
@@ -218,18 +164,9 @@ func (ssp *SimpleSpamPolicy) PostBlockAccept(tx abci.Tx) (bool, error) {
 		blockCommands += count
 	}
 
-	// if too many votes in total - reject and update counters
+	// if too many votes in total - reject
 	if epochCommands+blockCommands >= ssp.maxAllowedCommands {
-		// update vote stats for the epoch
-		if partyRejectStats, ok := ssp.partyBlockRejects[party]; ok {
-			partyRejectStats.add(true)
-		} else {
-			ssp.partyBlockRejects[party] = &blockRejectInfo{total: 1, rejected: 1}
-		}
-		if ssp.log.GetLevel() <= logging.DebugLevel {
-			ssp.log.Debug("Spam post: party has already submitted the max amount of commands for "+ssp.policyName, logging.String("txHash", hex.EncodeToString(tx.Hash())), logging.String("party", party))
-		}
-		return false, ssp.tooManyCommands
+		return ssp.tooManyCommands
 	}
 
 	// update block counters
@@ -238,30 +175,19 @@ func (ssp *SimpleSpamPolicy) PostBlockAccept(tx abci.Tx) (bool, error) {
 	}
 	ssp.blockPartyToCount[party]++
 
-	// update party and block stats
-	if partyRejectStats, ok := ssp.partyBlockRejects[party]; ok {
-		partyRejectStats.add(false)
-	} else {
-		ssp.partyBlockRejects[party] = &blockRejectInfo{total: 1, rejected: 0}
-	}
-	return true, nil
+	return nil
+}
+
+func (ssp *SimpleSpamPolicy) RollbackProposal() {
+	ssp.blockPartyToCount = map[string]uint64{}
 }
 
 // PreBlockAccept checks if the commands violates spam rules based on the information we had about the number of existing commands preceding the current block.
-func (ssp *SimpleSpamPolicy) PreBlockAccept(tx abci.Tx) (bool, error) {
+func (ssp *SimpleSpamPolicy) PreBlockAccept(tx abci.Tx) error {
 	party := tx.Party()
 
 	ssp.lock.RLock()
 	defer ssp.lock.RUnlock()
-
-	// check if the party is banned
-	until, ok := ssp.bannedParties[party]
-	if ok {
-		if ssp.log.GetLevel() <= logging.DebugLevel {
-			ssp.log.Debug("Spam pre: party is banned from "+ssp.policyName, logging.String("txHash", hex.EncodeToString(tx.Hash())), logging.String("party", party))
-		}
-		return false, ssp.banErr(time.Unix(0, until).UTC())
-	}
 
 	// check if the party has enough balance to submit commands
 	balance, err := ssp.accounts.GetAvailableBalance(party)
@@ -269,7 +195,7 @@ func (ssp *SimpleSpamPolicy) PreBlockAccept(tx abci.Tx) (bool, error) {
 		if ssp.log.GetLevel() <= logging.DebugLevel {
 			ssp.log.Debug("Spam pre: party has insufficient balance for "+ssp.policyName, logging.String("txHash", hex.EncodeToString(tx.Hash())), logging.String("party", party), logging.String("balance", num.UintToString(balance)))
 		}
-		return false, ssp.insufficientTokensErr
+		return ssp.insufficientTokensErr
 	}
 
 	// Check we have not exceeded our command limit for this given party in this epoch
@@ -277,10 +203,10 @@ func (ssp *SimpleSpamPolicy) PreBlockAccept(tx abci.Tx) (bool, error) {
 		if ssp.log.GetLevel() <= logging.DebugLevel {
 			ssp.log.Debug("Spam pre: party has already submitted the max amount of commands for "+ssp.policyName, logging.String("txHash", hex.EncodeToString(tx.Hash())), logging.String("party", party), logging.Uint64("count", commandCount), logging.Uint64("maxAllowed", ssp.maxAllowedCommands))
 		}
-		return false, ssp.tooManyCommands
+		return ssp.tooManyCommands
 	}
 
-	return true, nil
+	return nil
 }
 
 func (ssp *SimpleSpamPolicy) GetSpamStats(party string) *protoapi.SpamStatistic {
@@ -289,7 +215,6 @@ func (ssp *SimpleSpamPolicy) GetSpamStats(party string) *protoapi.SpamStatistic 
 	return &protoapi.SpamStatistic{
 		CountForEpoch:     ssp.partyToCount[party],
 		MaxForEpoch:       ssp.maxAllowedCommands,
-		BannedUntil:       parseBannedUntil(ssp.bannedParties[party]),
 		MinTokensRequired: ssp.minTokensRequired.String(),
 	}
 }