+++ core/validators/topology_snapshot_test.go
@@ -18,6 +18,7 @@ package validators_test
 import (
 	"bytes"
 	"context"
+	"encoding/hex"
 	"testing"
 
 	"code.vegaprotocol.io/vega/core/types"
@@ -30,8 +31,6 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	abcitypes "github.com/tendermint/tendermint/abci/types"
-	types1 "github.com/tendermint/tendermint/proto/tendermint/types"
 )
 
 var topKey = (&types.PayloadTopology{}).Key()
@@ -164,10 +163,8 @@ func TestTopologySnapshot(t *testing.T) {
 
 func updateValidatorPerformanceToNonDefaultState(t *testing.T, top *validators.Topology) {
 	t.Helper()
-	req1 := abcitypes.RequestBeginBlock{Header: types1.Header{ProposerAddress: address1, Height: int64(1)}}
-	top.BeginBlock(context.Background(), req1)
+	top.BeginBlock(context.Background(), 1, hex.EncodeToString(address1))
 
 	// expecting address1 to propose but got address3
-	req2 := abcitypes.RequestBeginBlock{Header: types1.Header{ProposerAddress: address3, Height: int64(1)}}
-	top.BeginBlock(context.Background(), req2)
+	top.BeginBlock(context.Background(), 1, hex.EncodeToString(address3))
 }